refactor : 패키지 구조 변경 각 클래스 별 역할에 따라 패키지로 묶어 관리하도록 수정했다.
