+++ src/main/java/model/Cookie.java
@@ -1,4 +1,6 @@
-package webserver;
+package model;
+
+import webserver.HTTPExceptions;
 
 import java.util.HashMap;
 import java.util.Map;
