+++ src/main/java/webserver/request/staticResource/StaticResourceHandler.java
@@ -1,7 +1,14 @@
 package webserver.request.staticResource;
 
+import model.ContentType;
+import model.Cookie;
 import webserver.*;
+import webserver.request.HTTPRequestBody;
+import webserver.request.HTTPRequestHeader;
 import webserver.request.RequestProcessor;
+import webserver.response.HTTPResponse;
+import webserver.response.HTTPResponseBody;
+import webserver.response.HTTPResponseHeader;
 
 import java.io.File;
 import java.io.IOException;
@@ -24,7 +31,7 @@ public class StaticResourceHandler implements RequestProcessor {
             responseBody = new HTTPResponseBody(Files.readAllBytes(file.toPath()));
 
             responseHeader.setStatusCode(200);
-            responseHeader.addHeader("Content-Type", ContentTypeMapper.getContentType(path));
+            responseHeader.addHeader("Content-Type", ContentType.getContentType(path));
             responseHeader.addHeader("Content-Length", Integer.toString(responseBody.getBodyLength()));
         } catch (HTTPExceptions e) {
             responseHeader.setStatusCode(e.getStatusCode());
