+++ src/main/java/webserver/response/HTTPResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 public class HTTPResponse {
     private HTTPResponseHeader responseHeader;
