+++ src/test/java/webserver/HTTPRequestHeaderTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.request.HTTPRequestHeader;
 
 import static org.junit.jupiter.api.Assertions.*;
